test: Users 추가로 인한 테스트 수정
